Improve the CLI and add image resizing to median_cut
Handling of files in the CLI has changed so if the last file does not
exist it is assumed to be the output file. An explicit argument for the
output file has also been added. The image passed to median_cut is
resized to reduce the amount of work that needs to be done.